Move leaderboard to it's own file
